add pypsa loader to load a given pypsa network with its timeseries properties
This adds a new loader for arbitrary pypsa networks.

Examples are given from the pypsa examples as:
* `ac_dc_meshed`
* `scigrid_de`
* `storage_hvdc`

These only contain a few hours of simulations, as the corresponding snapshots are not that large.
Besides of the `loads` and `generators`, the buses are loaded and used for the grid visulaization and calculation, if a redispatch or grid related market is used.