Fix broken pypi builds (#407)
<!--
SPDX-FileCopyrightText: ASSUME Developers

SPDX-License-Identifier: AGPL-3.0-or-later
-->

# Pull Request

## Related Issue
Closes #[issue number] (if applicable)

## Description
The switch from poetry to setuptools has introduced a critical error
into the building of the package and the latest release of ASSUME on
pypi is broken as no files of the framework are provided and it doesn't
work. This commit should fix this issue. The pypi release has also been
updated using this update to ensure the functionality of the package. It
also fixes the problem of importing steel_plant when no pyomo is
installed.

## Changes Proposed
- use [tool.setuptools]
packages = {find = {}}
- use units directly inside __init__.py for unit types to adress issues
with pyomo
- fix smaller pandas things

## Testing
Created a new environment and ran the examples and notebooks

## Checklist
Please check all applicable items:

- [x] Code changes are sufficiently documented (docstrings, inline
comments, `doc` folder updates)
- [ ] New unit tests added for new features or bug fixes
- [x] Existing tests pass with the changes
- [ ] Reinforcement learning examples are operational (for DRL-related
changes)
- [x] Code tested with both local and Docker databases
- [x] Code follows project style guidelines and best practices
- [x] Changes are backwards compatible, or deprecation notices added
- [x] New dependencies added to `pyproject.toml`
- [ ] A note for the release notes `doc/release_notes.rst` of the
upcoming release is included
- [x] Consent to release this PR's code under the GNU Affero General
Public License v3.0

## Additional Notes (if applicable)
[Any additional information, concerns, or areas you want reviewers to
focus on]

## Screenshots (if applicable)
[Add screenshots to demonstrate visual changes]