Remove the profile learning strategy (#566)
This pull request removes the existing profile learning strategy from
the framework. The current implementation of this strategy has been
identified as adding unnecessary complexity and weight to the system, as
it only bids a single price for the entire day, which does not align
with the desired functionality. The profile learning strategy will be
reevaluated and replaced with a new "new day ahead" strategy in the
future.