set accepted_price of rejected bids to zero
This value should never be needed.
When calculating anything on the bids, one should only consider the accepted ones (e.g. accepted_volume != 0) the price is only given to have an accepted price at all, but it does not make any sense, as the bid is not accepted..
Using the market clearing price - or the first market_meta price for all bids does not make sense here.