Have a single forecaster per unit instead of one global forecaster (#660)
The global forecaster provides a way to store the input data of the units in a centralized dict.
However, this approach complicates the identification of misconfiguration and may result in simulations that run with zero values producing wrong outputs, making it difficult for users to identify the problem.

Assigning each units its own forecaster solves these issues by requiring necessary fields to be set and simplifying validations of fields on unit-level. This also solves issue #635, where the resolved forecaster field depends on the id of a unit.