tests: utcoffset of timestamp(0) did fail on windows
This is fixed by calculating the utcoffset in 2020 instead of beginning of time, so that all tests pass.
It is still unclear, why the utcoffset fails on windows for that, but it isn't a problem anyway, as the date to calculate the offset from in the tests is arbitrary anyhow.