tests: utcoffset of timestamp(0) did fail on windows (#671)
This is fixed by calculating the utcoffset in 2020 instead of beginning
of time, so that all tests pass. It is still unclear, why the utcoffset
fails on windows for that, but it isn't a problem anyway, as the date to
calculate the offset from in the tests is arbitrary anyhow.

## Checklist
- [ ] Documentation updated (docstrings, READMEs, user guides, inline
comments, `doc` folder updates etc.)
- [ ] New unit/integration tests added (if applicable)
- [x] Changes noted in release notes (if any)
- [x] Consent to release this PR's code under the GNU Affero General
Public License v3.0

## Additional Notes (optional)
<!-- Anything the reviewer should pay special attention to, known
limitations, rollout plan, etc. -->